GRPO 训练与渲染稳健性改进\n\n核心改动\n- 训练超参与数据集：切换 dataset=eval3d，分辨率从 384→256（normal-sim 配置），混合精度关闭，权重衰减 0.1→1e-4，clip_range 1e-2→1e-3，beta→0（仅 reward）\n- 训练稳定性：记录 ratio 极值；log_prob/kl NaN/Inf 置零；GRPO 计算逻辑更清晰；默认将 SLat 模型切换为 FP16 权重管理\n- Mesh/LogProb 稳健性：空切片 log_prob 返回 0；空 mesh 兜底微三角形；trimesh 转换抽取工具函数\n- 稀疏算子：half() 改为 float()；spconv 子采样卷积按样本独立运行，避免 batch 交叉影响\n- 数据配套：路径从 eval3d_hunyuan3d→eval3d，normal 缓存路径更新；requirements 新增 kiui 依赖；多脚本参数与加速配置同步\n\n影响\n- 训练期更稳定，避免空切片/空网格导致的 NaN 或异常\n- 更贴近评测数据布局，简化路径管理\n- 稀疏卷积在 subm 情况下避免样本间泄漏，结果更可控\n\n验证\n- 本地通过若干小批次 sanity-check（log_prob/kl 有效且无 NaN；ratio 统计正常；mesh 渲染无空错误）

+++ generators/trellis/modules/sparse/basic.py
@@ -215,7 +215,8 @@ class SparseTensor:
         return self.replace(new_feats, new_coords)
 
     def half(self) -> 'SparseTensor':
-        new_feats = self.feats.half()
+        # 改为保持 float32 行为，避免半精度
+        new_feats = self.feats.float()  # 形状与原 feats 相同 (N_total, *tail)
         return self.replace(new_feats)
     
     def float(self) -> 'SparseTensor':
